- bug fix: forced conversion of uint64_t to unsigned long int to prevent a warning; - bug fix: some router variables (for power measurement) were not properly initialized.

+++ src/noc/orion/test_permu.cc
@@ -126,7 +126,7 @@ int main(int argc, char **argv)
 		printf("%d\n", d_tab[in_p]);
 		cnt += d_tab[in_p];
 	}
-	printf("total: %lu\n", cnt);
+	printf("total: %lu\n", (unsigned long int) cnt);
 
 	exit(0);
 }
